+++ automation/src/epilot/models/shared/triggerworkflowcondition.py
@@ -5,7 +5,13 @@ import dataclasses
 from ..shared import comparison as shared_comparison
 from dataclasses_json import Undefined, dataclass_json
 from epilot import utils
-from typing import Any, Optional
+from typing import Optional, Union
+
+
+
+@dataclasses.dataclass
+class TriggerWorkflowConditionValue:
+    pass
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -15,6 +21,6 @@ class TriggerWorkflowCondition:
     comparison: shared_comparison.Comparison = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('comparison') }})
     schema: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('schema') }})
     source: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('source') }})
-    value: Optional[Any] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('value'), 'exclude': lambda f: f is None }})
+    value: Optional[Union[str, float, list[str], list[float]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('value'), 'exclude': lambda f: f is None }})